[kernel][hypervisor] Make hypervisor log message provenance clearer
Some log messages emitted by the hypervisor, when read out of context,
have an unclear meaning. For example, the kernel log message:

  Unhandled rdmsr 0x3a

is emitted during the normal boot of a Linux guest. Looking at the logs,
though, it may appear that the host is having problems.

This CL updates log messages in the hypervisor to be clearer that they
are being emitted by the hypervisor, and that the cause of the issue is
the guest:

  hypervisor: Unhandled guest RDMSR 0x3a

Change-Id: I3a3b796a0ee7dbc23cd92a872a7adea2753e0470
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592832
Commit-Queue: David Greenaway <dgreenaway@google.com>
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>