[zstd] Optimize unconditionally
Prior to this change zstd was optimized for speed on the host
irrespective of the global optimization settings in order to reduce the
build-time cost of zstd compression. This extends that to the target as
well in order to reduce the resolve-time cost of zstd compression.
Measured in qemu-x64 with KVM this gives a ~70% increase in throughput:

Before:
  $ time fx shell pkgctl resolve fuchsia-pkg://fuchsia.com/debian_guest
  resolving fuchsia-pkg://fuchsia.com/debian_guest with the selectors []
  Error: Failed to resolve fuchsia-pkg://fuchsia.com/debian_guest
  Caused by:
      there is no space available to store the package or metadata
  real    3m53.166s
  user    0m0.871s
  sys     0m0.179s

After:
  $ time fx shell pkgctl resolve fuchsia-pkg://fuchsia.com/debian_guest
  resolving fuchsia-pkg://fuchsia.com/debian_guest with the selectors []
  Error: Failed to resolve fuchsia-pkg://fuchsia.com/debian_guest
  Caused by:
      there is no space available to store the package or metadata
  real    2m18.137s
  user    0m0.817s
  sys     0m0.251s

Bug: 65962
Change-Id: If846c2abcc415d1320c0b390570b60e8942d0fc7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592869
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Brett Wilson <brettw@google.com>