[plasa] Package the plasa artifacts with the SDK
With this change, the plasa fragment files ('*.fragment.plasa.json') are
now packaged with the SDK.

The fragment files for now contain the minimum information for the API
reports to be usable in test coverage reporting.

Also, the fragment generation is implemented as a proof of concept only
in `sdk_source_set`, with other sdk source targets coming soon.

Tested: Manually using the `generate_plasa_artifacts = true` and
  `build_sdk_archives`, then manually inspecting the resulting
  archives.
Bug: 82512
Change-Id: If4fa2b17f10dd1baf8f0def95db35eb5a1d54714
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589623
Reviewed-by: Jerry Belton <jcecil@google.com>
Reviewed-by: Misha Gridnev <gridman@google.com>
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>