[packet-formats] Refactor TCP header storage
Add the `DataOffsetReservedFlags` type, which stores the Data Offset
field, the reserved zero bits, and the flags. This gives us a simpler
mechanism to copy all of the flags into a new `TcpSegmentBuilder` in the
`TcpSegment::builder` method. It also prepares us for
https://fxrev.dev/I0c8266167ab3bc05885346bdeb6c18b502b93be3, which adds
the [`TcpSegmentRaw::builder`] method.

While we're here, add a getter for the PSH flag, add some documentation
here and in `ipv4` and `ipv6`, and fix a typo in
`ipv4::tests::new_builder`.

Change-Id: I9060cce241c6946468f1d043223895d283da2f00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590743
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>