[ffx] Schema from deserialize
Create a library which can construct a serializable witness to a type T
if T: Deserialize. This uses a Deserializer to keep track of the
different parts of T as it is driven by the deserialize implementation
that is implemented/derived for T.

Enums are a bit tricky with this approach, but I worked out a way to
solve those issues. I still need more tests.

The goal is to call this for the output type of a plugin that is
specified as an attribute to the ffx_plugin procedural macro. The result
can then be checked in and compared with subsequent changes to ensure a
consistent schema.

Change-Id: I85b2915f0390198789b84597476adc2aa6b86ad9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576565
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>