[ffx] Add self-tests for component list
This is basically a smoke test that things more or less work. It brought
out a bug in the self-test runner which is fixed in the ancestor to
this. I plan to add more tests here but this is getting the groundwork
in place.

Bug: 82378

Multiply: ffx-e2e-with-target

Change-Id: Ia71c847ceac5ea58a960ab172e1e2b16a9e8b585
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585291
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>