[kernel][x86] Avoid CPU init stack overflows in kasan/lockdep builds.
The x86 arch sets up the kernel stacks for CPU init threads in a special
way, such that the stack pointer is initialized while transitioning into
long mode. Once in long mode, the init thread continues to use the same
stack location for the rest of the init. However, prior to this change,
the stack pointer was initialized to only one page above the stack base,
regardless of configured size for kernel stacks. This can result in a
stack overflow when features such as kasan and lockdep are used
together.

This CL changes the CPU init thread setup to pass the actual top of the
kernel stack to the low-level CPU code, such that the configured stack
size is correctly used and maintained into the higher-level the init.

Fixed: 86361
Test: fx set bringup.x64 --variant=kasan --args 'lock_dep_enabled=true'
boots.

Change-Id: Ic19d7237e7922a0f6cc3b4286008b0d51dd085ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592637
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Corey Tabaka <eieio@google.com>