[fx] Add ffx device discovery to `fx shell`
fxb/85247 describes a shortcoming in the ffx emu workflow, where certain
fx tools couldn't communicate with the emulator instance unless the user
manually indicated the address and port of the target device using `fx
set-device`. We tracked this down to the fact that the `fx shell` tool
calls a library function which only checks environment variables and
hidden files for device information, and doesn't check the ffx daemon.

This CL introduces a new library function to the fx "vars" file,
get-fuchsia-device-port. This function starts by calling the original
ssh port function, get-device-ssh-port, then falls back to ffx if it
doesn't find one. We then modify `fx shell` to use this function instead
of the original, which waterfalls to other tools that call shell
directly (such as debug and fidlcat). This removes the requirement for
the user to call `fx set-device` for these tools, even when running on
SLIRP instead of TUN/TAP.

There are about half a dozen other tools which still call the old
function directly, and these can be migrated in future changes if deemed
appropriate.

Change-Id: Ib422cb4fad5db2775f7bd8b70da8f8470fbc30db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591990
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Gary Boone <gboone@google.com>