[component_manager] Don't auto-bind to ancestors
This change removes behavior that automatically binds to a component's
ancestors when it is bound to. This was meant to support a principle
that parents should run with their children, but this is no longer a
guarantee we aim to provide.

Also, remove duplicate tests in binding.rs (that were rehomed to
lifecycle.rs).

Change-Id: Id5ed6a6f85e399a4f2fc31896c3dba2b945e8646
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589471
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Commit-Queue: Gary Bressler <geb@google.com>