[clippy] update script args/usage
Replacing `--files-to-targets` with `--get-outputs` should be more
future proof because when we want to enable clippy targets for host and
other toolchains, we'll need to build a list of output files rather
than a list of targets.

Also introduce the `--all` flag to get the list of, or run clippy on,
every clippy target in the current build configuration.

Also remove some #![deny(warnings)] from the codebase. These annotations
are redundant as we apply it tree wide, and it interferes with clippy
because its lints are produced as warnings that we don't want to deny.

Bug: 86506

Change-Id: Icc434be84d26bea80200338b939239287c857093
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593244
Commit-Queue: Joseph Ryan <josephry@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>