[build][clippy] Add clippy script to action_tracer ignore list
The rust compiler performs some reads of files that aren't truly
dependencies of a given target. For example it will read multiple
libraries in it's library search dirs and pick the correct one to
actually use. This makes it fail our hermeticity checker because those
extra file reads show up as spurious dependencies.

Because we only run the tracer on GN actions, we haven't had to address
this with rustc yet, but clippy (which is a thin layer ontop of rustc) is
implemented as an action, so we need to whitelist it here.

Change-Id: I250f292d9a1a1ee1e755fefb6004e7ef410a9fee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592632
Commit-Queue: Joseph Ryan <josephry@google.com>
Reviewed-by: David Fang <fangism@google.com>