[plasa] Another internal header in DDK needs exclusion
The added header is not intended for direct inclusion by
the user code, and will error out in that situation.  Add it
to the list of non-public headers so that the Plasa API stub does
not attempt to use it.

Change-Id: I4f11d9098bbcb9e2dfb9220fb768bc5d5170c9a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592603
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeremy Manson <jeremymanson@google.com>