[chromium] Allow setting a package name in generated targets
The target name for the browser is `chrome_browser` but the package name
is just `chrome`. Under the assumption that we want to retain the former
for clarity in the build targets, this adds a package_name parameter to
the GN templates. The parameter mimics the existing prebuilt package
templates in //build/packages.

Bug: 86236

Change-Id: Id88967be704bded2bd6e3613f3fefcebbb10ecf8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592830
Reviewed-by: Chong Gu <chonggu@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Alex Legg <alexlegg@google.com>