[testrunner] Use a struct to represent flags
Use a local variable struct instead of a collection of global variables.
When all the flags were global variables it was easy to accidentally use
a global variable within a function that was intended to be stateless,
making it harder to test.

Change-Id: Ia279599b3542edd39ef8e62b90af38fd5a8a5672
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593703
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>