[starnix] Add explicit dependency on wayland bridge
Some configurations (e.g., core) do not include wayland (or //src/proc),
but are used for development. Specifically, developers will add
`--with //src/proc` to their fx set. This includes starnix, but not
wayland. This surfaces as an error for a missing capability from the
wayland bridge (child) to starnix (parent).

CQ does not catch this because starnix is not built as part of core, and
on workstation wayland bridge is included.

This fixes the issue by adding an explicit dependency on the wayland
bridge package to starnix.

Change-Id: I0b6ca5ff370da4af50f2e4ba623c42fb07d62479
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593365
Reviewed-by: Steve Austin <steveaustin@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>