[ffx] Add default ssh port 8022 to ffx emu
go/ffx-set-device-decision explains why we should have a default port
set for `ffx emu` instead of always defaulting to a randomly generated
port number. This CL adds 8022 as that default value.

Additionally, while testing port-maps, we found that the short-form
port-map syntax (tcp:<host>:<guest>) was not recognized by ffx emu when
deciding if a default ssh port was needed or not. This CL adds a check
for short-form port-map entries before assigning the default or
allocating a dynamic port.

Bug: 85247

Change-Id: Ie65dbc29fbbeba46f7ba22ba3b6011406a5b6df8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592121
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>