[audio_core] Correctly handle stream first-packet
When a stream starts, it's first packet will not coincide with
the beginning of the MixJob. This "position discontinuity" is not
a problem, but we'll need to skip over some number of frames in
the Dest output buffer and advance our running position in Source
stream accordingly.

During mixing/resampling, two timelines describe the position
relationship between source stream and destination stream: the
"intended" relationship based on zx::clocks, and the "in reality"
relationship based on micro-SRC applied to eliminate position
error. Although over the long-run these timeline relationships
should be equivalent, they cannot be identical because we only
discover clock rate differences after-the-fact. For this reason,
when advancing source position, we must always advance by an
integral multiple of the resampler's "step_size +
rate_modulo/denominator" stride factor, which encodes this "in
reality" position relationship.

Previously, to estimate how many dest frames to skip over to
reach the beginning of a source stream's first packet, we used
TimelineFunction 'info.dest_frames_to_frac_source_frames'. This
seems as if it would be perfect for this purpose, but in fact it
is the "ideal" timeline that encodes the clock relationships
rather than the "in reality" one that instead includes the
micro-SRC being applied. Now, we calculate the dest frame advance
using step_size and rate_modulo/denominator. This allows us to
reenable a number of position-related CHECKs that ensure correct
state in the resamplers.

This problem was most reliably exposesd, and the solution
verified, using the new suite audio-core-fidelity-test.

Because the resamplers use positive-filter-length to produce a
proper ramp-in for its filter, it is not necessary to feed
frames of silence instead of simply advancing the position.
In the future when we better handle intentional and unintentional
client discontinuities, we may choose to feed silence upon client
underrun so that proper ring-out and ramp-in occurs in the case
of unintentional gaps between packets.

Bug: fxb/85359

Change-Id: I0f704921bf1106e0f2efc6c93d1a505a25e414d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589493
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>