[audio_core] V2 API for creating effects over FIDL
In audio effects V2, audio_core applies effects by calling the
fuchsia.audio.effects/Processor FIDL protocol. During initialization,
these Processor channels must be configured and handed to audio_core. In
practice, the channel may speak to an entirely different component that
is defined in a vendored directory, or even out-of-tree. This implies
that we can't hardcode channel configuration in audio_core -- we need
some way to import that configuration on a per-product basis.

Our solution is to create a new FIDL protocol, ProcessorFactory, which
has a single method, Create(name), which returns a Processor for the
effect with the given name. This protocol _may_ be transitional -- it
_may_ be deleted once we transition to vnext, where audio system
initialization will be entirely controlled by product-specific code.
Hence, for now, we leave the effect |name| completely opaque.

We also considered hiding this FIDL protocol behind a shared library.
This would allow us to keep the protocol out of the public SDK, at the
cost of some additional implementation complexity. See fxrev.dev/565962
for further discussion.

This CL defines ProcessorFactory, adds a simple C++ wrapper around the
protocol, and adds an implementation of the protocol for use in tests.

Bug: 80067
Change-Id: I3dd28b33abe3dc6942517cd085067583664bfe3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564802
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
API-Review: Dale Sather <dalesat@google.com>