[kernel][vm] Process the DontNeed queue
Process the DontNeed pager queue as well when processing LRU for page
queue rotation. This ensures that any pages in the DontNeed queue which
have been subsequently accessed, are fixed up and moved to the
corresponding pager queue. If the DontNeed queue is not processed when
advancing LRU, the queue indices in the vm_page_t can become old to the
extent of appearing to be valid again once the queues wrap, throwing off
queue counts. Currently the DontNeed queue is processed only when trying
to evict from it, which is not sufficient.

This CL introduces a pair of toggle DontNeed queues, and a generation
count to track which one is the current DontNeed queue. During
processing of the current queue, pages that do not require fixup are
moved to the toggle queue. This allows the processing thread to drop the
PageQueues lock between performing fixed amounts of work, and resume at
the tail each time until the current queue is empty, without having to
maintain a cursor which would be tricky with dropping the lock.

This fixes an existing bug where pages in the DontNeed queue which have
been accessed can result in incorrect active/inactive counts.

Test: kernel-unittests-test

Bug: 85768
Change-Id: If2ece74ab52faf8e0370fb59d262dfa5bf9ff826
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592153
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Fuchsia-Auto-Submit: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>