[starnix] fd_impl_nonblocking
Remove the default implementation of FileOps::wait_async
and create the fd_impl_nonblocking macro for a dummy
implementation of wait_async. This way, implementors
of FileOps will either need to mark the object as
non-blocking with this macro or implement wait_async
and future blocking-related methods.

Change-Id: Ic3a7460edc658fc390948067564d71202733b955
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594749
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Steve Austin <steveaustin@google.com>