[audio_core] Don't tune monotonic clocks
Detect when two clocks perfectly track CLOCK_MONOTONIC, and
do not actively synchronize them in that case.

When a device clock is in the monotonic domain and a custom
client clock is a clone of CLOCK_MONOTONIC that has not yet
been rate-adjusted, allow them to progress without taking
active synchronization measures. If/when we detect that the
client monotonic-clone HAS been updated (as revealed by the
generation_counter field), we will forever treat it as NOT
identical to CLOCK_MONOTONIC (and thus will rate-adjust it,
like any other clock).

This change improves audio quality because the resampler
stays position-locked rather than slightly "jittering" by
a few nanoseconds on any particular mix. This measurably
reduces distortion in the midrange for streams that do not
provide their own clock but use frame rates other than
48 khz. The coming test suite audio-core-fidelity-test
identifies the benefit (for 44.1->48 conversion) to be most
notable in the 500 hz to 12.5 khz range, where SiNAD is
improved by an average of 5.1 dB (most notably by 9.4 dB
at 2.5 khz).

fx test -o audio-libclock-unittests audio_core_unittests

Bug: fxb/86485

Change-Id: I03c90ab4fb873c2d7fea244c3e032063d95eb334
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592571
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>