[ffx] Fix e2e test flakes
I added this code to kill zombie daemons but it is too aggressive and is
causing flakes. This makes it not an error to kill a left over daemon
but still does the kill if necessary.

I also increased the per-test timeout to 12s because 10s is a race with
the default proxy connect timeout so the error messages are test case
timeouts usually instead of the real timeout that is causing the
problem. This should make those flakes provide more diagnostic
information.

Bug: 84938
Fixed: 86030

Multiply: ffx-e2e
Change-Id: I2440e91a18ac379ee8f1136f722a9110c9569d27
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594862
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>