[magma] Rename RegisterIo::Write32 to Write32Flipped
This is part 1 of flipping the order of the arguments to
RegisterIo::Write32. The order of the arguments to the RegisterIo::Hook
methods is also swapped for consistency.

This patch makes it easy to identify which callers are using the old
ordering. Part 2 will reorder the arguments everywhere and rename
everything back to the old names.

Bug: 86716
Change-Id: I7887d2a1af385a53de9ed8a69eecded1b3ef273c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594864
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Craig Stout <cstout@google.com>