[driver_manager] Start separating v1 code into a subdir
This change reorganizes code that is only used in DFv1. This v1 code
is largely in maintenance mode as we plan to flip over to v2 in the
future. Future changes will continue to split out existing classes to
make it easier for us to delete this code once we no longer need it.

Header includes were also updated to be style guide compliant.

Test: No functional changes
Change-Id: I165bc7ccaf210cf1fedfea70d364a3cd603a5ac1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594748
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>