[llcpp] Specify wire format version for device metadata
device_get_metadata returns a buffer containing FIDL bytes.
Currently, the wire format version for these bytes isn't
specified on decode. When the version is not specified, it
is assumed the initial bytes contain the FIDL header which
can be used to determine the version, but this is not the
case. This CL adds explicit naming of the wire format version.

Change-Id: I571d8e973d32db9315574fc871470d094bd44071
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594081
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>