[realm-builder] handle relative URLs at runtime
If a dynamic component is added to a realm with a relative URL,
component manager may construct an absolute URL on top of a
`realm-builder://` URL and then ask the realm builder server to resolve
it. We can detect this situation by looking for URL fragments (ex:
`#meta/foo.cm`) and then pulling up the requested manifest from the
test's package.

Fixed: 85884
Change-Id: Ia7533265dc089317ac232ca6ac58e3534a2e3737
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592964
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>