[ffx] Add golden tests for target list JSON
This change is to introduce tests that did not yet exist. There is no
functional change here. There is a bug in the json format that needs to
be fixed, but I wanted to get these tests of the current behavior in
place first and then update things in a follow-up.

Bug: 86588

Change-Id: I70305f1728cf24f5c264d354d0ad1294ca208648
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594022
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>