[build] Create drivers global documentation file
Generates a global documentation file with driver provided
data. This data is optionally included in a JSON file in driver's
folders. The generated global files is a JSON file to be used for
display on fuchsia.dev.

Deployment of the generated global file, validation the driver
provided JSON, and documentation are to be provided in subsequent
CLs prior to deployment.

A single driver provided data file is included for testing.

Bug: 81799

Change-Id: Idfff5e45c80bfdf9ffe9af590faaac9e7310b980
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593684
Commit-Queue: Andres Oportus <andresoportus@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>