[ffx] Add correct target_type to target list JSON
The JSON formatting code and the tabular formatting code use different
logic for computing the target_type field. I think this was just an
oversight when the tabular format was updated to take into account the
build configuration information, as the JSON code was not updated at
that time.

This unifies the logic for both so hopefully it is clear now if
something changes for one that it will/should change for the other.

Bug: 86588

Change-Id: I48e418261023480aaf3dd88635fb54e8e04dd229
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594023
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>