[driver_runtime] Add handle implementation.
This will be used for implementing fdf_channel_t.

This is closely based on the kernel handle table implementation.

Each process supports up to 64K handles. This is implemented
as a handles table directory which can hold 64 tables,
of which each table holds 1024 handles. We allocate 1 table at a time.

The fdf_handle_t value consists of:
 - the generation number: incremented each time the handle is reused
 - the directory index number: for locating the handle table
 - the index number: for locating the Handle in the handle table
 - reserved bits: 2 LSB, where LSB is 0. This differentiates it from a
   zircon handle which has both bits set to 1.

The directory index and index number allow for quick lookup of a
Handle given a fdf_handle_t.

Deleted handles are recycled onto free_handles_list_, which
can be used for new allocations. When reusing a handle, the
same directory index and index are preserved, (so that we
can correctly locate the corresponding Handle object),
but a new generation number is set to avoid returning a
recently deleted fdf_handle_t value.

Bug: 85577

Change-Id: Ia739c9c9eeed21eb4be3f89b17572752bfd5481c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593369
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>