[a11y] Return status from input.accessibility.PointerEventRegistry
This change modifies the semantics of
fuchsia.ui.input.accessibility.PointerEventRegistry to match
fuchisa.ui.policy.accessibility.PointerEventRegistry so that we can
later replace the policy namespace.

NOTE: This is NOT a breaking change. Root presenter was the only
implementer of this protocol, and we removed the implementation in
this change: https://fuchsia-review.googlesource.com/c/fuchsia/+/496699.

TEST: n/a
Bug: 86102
Change-Id: I06ab852a58bc2e396434ed380a711a184d422d6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592601
Reviewed-by: Alice Neels <neelsa@google.com>
API-Review: Alice Neels <neelsa@google.com>
Commit-Queue: Alexander Brusher <abrusher@google.com>