[blobfs] Use OP_DONT_NEED for inactive blobs
When blobs transition to being inactive, hint on their associated VMOs
that their pages can be evicted under memory pressure. Currently, this
is done under the covers with a kernel hack that performs the same
operation when no clones remain for a VMO, which is when blobfs
considers a blob inactive. The kernel hack is going away now that the
proper hinting API is available for use.

Bug: 65334
Change-Id: I7fabbfe14ee0330ae8153819b47fefcdfa951279
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594978
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>