[third_party] Update generic-array from clean
It seems that someone decided to check in DOS line endings (CRLF)
instead.  This messes up our change detection and we get this
request to add line endings everywhere.

Perhaps it makes sense to just take the change in to simplify
further imports - that is if it does not complicate other things.

Change-Id: Ic5177e48bda621766141841d9e2b5030d20e8a26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595390
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>