[fidlc] Eliminate special case for attribute validation on protocols
This CL refactors attribute constraint validation to make it more
uniform. In particular, attributes on protocols were previously treated
as a special case and validated on each of the protocol's methods
instead of on the protocol itself. This is no longer the case: instead,
the validation methods iterate over methods internally.

This also fixes a bug where @max_bytes and @max_handles were allowed on
protocols and methods but were not validated.

Change-Id: I4d63e6fc4b02b4cebbb1a2919c99e6122600a11c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593743
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>