[fidlgen_cpp] Reduce usage of MethodResult
**Overall strategy**

To correct the implementation of the error syntax when using externally
composed methods, we must reduce our reliance on "all the information"
for the error syntax.

Currently, we incorrectly assume that we can look up various definitions
(the result union, the success struct, the error type). Then, we arrange
all this information in a `MethodResult` value that we attach to the
method using the error syntax. We also attach this object to the union
which is used as the result union.

With externally composed methods, none of that can work, and we must
therefore correct this incorrect assumption, and adapt all backends to
operate with less information. At the same time, some extra information
will need to be provided by `fidlc` to compensate.

**This specific change**

This is the first of a series of changes which reduces the reliance of
backends on the `MethodResult` fields which we cannot provide in the
externally composed case, i.e.:

* `ResultUnion` points to a union, which is not available with an
  external library
* `ValueMember`
* `ValueStruct`
* `ErrorMember`.

In this specific change, we remove the C++ backends use of
`ResultUnion`, and instead shift these usages to `ResultTypeIdentifier`
which is simply a fully qualified identifier of the result type.

In subsequent changes, more usages will be removed, and eventually the
JSON IR will be changed to provide the fields backends need directly,
thus making externally composed methods work similarly then other
methods.

On the Go cleanup front, we change all `fidlgen_cpp` declaration types
to implement the `Kinded` interface on pointer receivers, rather than
value receivers. Because we mutate these values in various places, using
value types is very error prone. By changing these to pointer types we
avoid a whole class of bugs. This also follows the advice of the Go
rubric
https://fuchsia.dev/fuchsia-src/concepts/api/go#implementing-interfaces.

Bug: 64992
Change-Id: Idaa3a758e53cca8bc54282a50be55cf07e54ccf4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591981
Commit-Queue: Pascal-Louis Perez <pascallouis@google.com>
Fuchsia-Auto-Submit: Pascal-Louis Perez <pascallouis@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>