[fidlc] Remove name from AttributeArgSchema
This CL fixes the names of attribute arguments, starting by removing the
name field from AttributeArgSchema. Just as an AttributeSchema does not
know its own name, AttributeArgSchema should not either. Having two
copies of the name (in the std::map and in the value) led to bugs where
the std::map key was different from the AttributeArgSchema's name.

In addition, this CL:

* Defines a constant for "value" so it only appears in the code once.

* Makes the error message say "its required argument" rather than "its
  required 'foo' argument" when there is only one, as the latter
  suggests that @attr(foo="...") is desired, but that is forbidden.

* Fixes the test suite name in generated_name_tests.cc.

Change-Id: I50d4735aa9fdb9ff0eff1114eef27c5912593b2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595265
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>