[build][kernel] Simplify specifying kernel command line args
In the current build system, if a developer wants to specify an ad-hoc
kernel command line argument for local testing purpose, they have to:

1. Create a new directory, `local`, in their Fuchsia root directory.

2. Create a new file in this directory, `BUILD.gn` which contains
   a valid `kernel_cmdline` target, in turn containing a list of
   strings to add to the kernel command line.

3. Update the `dev_bootfs_labels` argument in their `args.gn` file
   to point to this new rule.

If the user has multiple builds and wants different sets of kernel
command line parameters for each build, the `local/BUILD.gn` file needs
to contain multiple different kernel_cmdline rules: one for each target.

This makes it hard to tweak kernel command line parameters for the
purposes of local testing and debugging.

This CL a new GN argument, `dev_kernel_cmdline`, which allows a
kernel command line argument to be specified by just adding it as
a string:

  fx set ... --args='dev_kernel_cmdline=["my_option=value"]'

Values in `dev_kernel_cmdline` are not expected to be checked in, but
rather are for local development only.

Change-Id: If469ee07d0fdbbb9e48ad84463ecce9d65f17940
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594705
Reviewed-by: Sarah Pham <smpham@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>