[fidlc] Add ForEachDeclMember
This CL adds a generic facility for applying a function to each
Attributable* member of a Decl*, and rewrites VerifyDeclAttributes in
terms of it. This reduces duplication and the chance of mistakes (in
fact there was a mistake: table members were validating placement twice
instead of placement and constraints). I also plan to reuse this for
FIDL versioning.

Test: fx test fidl-compiler
Change-Id: I96aa4218844ae6f0a1df6ee6923e3738b79de277
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593744
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>